requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.7.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.10.0
